Split Ship into ShipTakeDamage and ShipMove
