+++ Assets/Scenes/Scene.unity
@@ -1841,6 +1841,7 @@ GameObject:
   - component: {fileID: 349337797}
   - component: {fileID: 349337802}
   - component: {fileID: 349337801}
+  - component: {fileID: 349337803}
   m_Layer: 8
   m_Name: ship
   m_TagString: Player
@@ -1906,7 +1907,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: c1ac2cf42086c1a4cb995dfd14b04fe5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  explosion: {fileID: 1000013462590818, guid: ecdd96beb7f93494880c052dfccf9dc1, type: 2}
   speed: 10
   exSpace: 0.5
 --- !u!50 &349337798
@@ -2017,6 +2017,18 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   shield: {fileID: 1644133065064688, guid: d243e0883e8841348a99d5997c62233b, type: 2}
   Active: []
+--- !u!114 &349337803
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 349337794}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c2b0d6b5a4e5a3444ab527b564b9168c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  explosion: {fileID: 1000013462590818, guid: ecdd96beb7f93494880c052dfccf9dc1, type: 2}
 --- !u!1 &356110926
 GameObject:
   m_ObjectHideFlags: 0
@@ -10917,7 +10929,7 @@ RectTransform:
   m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 1680000587}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: -4478, y: -1516, z: 0}
+  m_LocalPosition: {x: -4540, y: -745, z: 0}
   m_LocalScale: {x: 1.0005332, y: 1.0005332, z: 1.0005332}
   m_Children:
   - {fileID: 646057972}
@@ -10928,7 +10940,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 1, y: 1}
-  m_AnchoredPosition: {x: -4478, y: -1516}
+  m_AnchoredPosition: {x: -4540, y: -745}
   m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &1680000593
