--- Assets/Scripts/Misc/CameraMoveWithShip.cs
@@ -1,20 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class CameraMoveWithShip : MonoBehaviour {
-
-    // Camera movement
-
-
-
-    // Use this for initialization
-    void Start () {
-		
-	}
-	
-	// Update is called once per frame
-	void Update () {
-     
-    }
-}
