+++ Assets/Scripts/Player/Ship/ShipTakeDamage.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: b9162431f2bb9144194a8843e8130776
-timeCreated: 1526818220
+guid: c2b0d6b5a4e5a3444ab527b564b9168c
+timeCreated: 1526926286
 licenseType: Pro
 MonoImporter:
   externalObjects: {}
